Refactoring and adjusted notify endpoint
